Fixed several compilation warnings

+++ src/include/widgets.h
@@ -295,7 +295,7 @@ public:
 
 	virtual int getNumberOfElements() override
 	{
-		return list.size();
+		return static_cast<int>(list.size());
 	}
 
 	virtual std::string getElementAt(int i) override
